requirements.txt
openai==0.27.8
langchain==0.0.196
pinecone-client==2.2.2
notion-client==2.0.0
tiktoken==0.4.0
python-dotenv==1.0.0
pypdf==3.9.1
